+++ NEWS.md
@@ -1,5 +1,9 @@
 # ggeffects 0.2.1
 
+## General
+
+* Support for `stanreg` models (pkg _rstanarm_).
+
 ## Bug fixes
 
 * Plotting raw data with `plot()` did not work for predictions at specific values (i.e. when certain levels of predictor where selected in square brackets).